# Optimize `ODataPath` operations