… src/Microsoft.OData.Core/ODataWriterCore.cs
… src/Microsoft.OData.Core/UriParser/SemanticAst/KeySegment.cs
… src/Microsoft.OData.Core/UriParser/SemanticAst/ODataPath.cs
… src/Microsoft.OData.Core/UriParser/SemanticAst/ODataPathExtensions.cs
… src/Microsoft.OData.Core/UriParser/Visitors/SplitEndingSegmentOfTypeHandler.cs
… src/Microsoft.OData.Edm/ExtensionMethods/ExtensionMethods.cs