<!-- markdownlint-disable MD002 MD041 -->

### Issues

*This pull request fixes #2814 .*

### Description

This PR does bunch of optimizations in different places to improve the performance of `ODataPath` operations. My focus was on `AppendKeySegment` since it took 7+% CPU on my tests, but the improvements affect other methods as well. Below is summary of the changes made.

- **`ODataPath`**
  - Reduced array copies and unnecessary array traversals (e.g. to check for null segments)
  - created a constructor that takes a `List<ODataPathSegment>` and sets it as the `path.Segments` property without performing a copy of checking for nulls. This is used in internal methods that create a new path based on a previous path by either adding or removing segments to a new segments list. In such cases, there's no reason to check whether we have null segments in the list since the old `ODataPath` was already valid, we also don't need to copy the segments list in the constructor since the `ODataPath` is being created from a newly created temporary collection, we can just use that list as is.
  - Replaced the use of `SplitEndingSegmentOfTypeHandler` for splitting a path with a basic while loop. Removed `SplitEndingSegmentTypeHandler` since it's no longer used.
- **`KeySegment` and retrieval of key values**
  - Like in the previous case, reduced array copies and validations
  - Added an internal constructor that accepts an `IList<KeyValuePair<string, object>>` and stores this list directly in the `ReadOnlyCollection` instead of calling `ToList()` on the `IEnumerable<KeyValuePair<string, object>>`. This constructor also optionally skips checking whether the key segment type matches the navigation source type
  - Reduced use of LINQ. Using a concrete List<T> to create the key value collection instead of `yield return`.  This avoid the use of `ToArray()` which was expensive. As a result I had to change the signature of the keys field from `KeyValuePair<string, object>[]` to `IList<KeyValuePair<string, object>>[]`
- **`as IEdmEntityType`casts**
  - reduced the number of casts happening. Frequent casts add up and we have a complex type of hierarchy. In some cases we would have a method call `Scope.ResourceType as IEdmEntityType` and then call another method which also performs the same cast. In such cases I pass the already cast `IEdmEntityType` as a parameter to the callee so that I can avoid performing the same casts in called methods.
- **`FullName()` Edm extension method**
  - This method is called frequently. If the element implements `IEdmFullNamedElement`, then the underlying `FullName()` method is called, otherwise it constructs the full name from combining `Namespace` and `Name` properties. The method is called most of the time with an element that implements `IEdmFullNamedElement` but it reads the `Name` and `Namespace` properties first. Reading those properties is somewhat expensive. So I optimized this method by moving the check for `IEdmFullNamedElement` to the top, before checking the `Name` and `Namespace` properties.
  - I reduced how often this method is called. There are many cases where the full name is passed as a parameter to methods when it's only used to generate an exception message when an error is thrown, but not used otherwise. So now instead of calling `FullName()` up the call stack, I just pass the entity type down an call `FullName()` only when the exception is thrown.

### Benchmarks

#### CPU Profiler

CPU dropped from > 7% to < 3% for the `AppendKeySegment` method

**Before**
![image](https://github.com/OData/odata.net/assets/8460169/bb08ef11-a30c-4127-ba82-c772845d609d)


**After**
![image](https://github.com/OData/odata.net/assets/8460169/aff0d6a4-ad7f-473b-a959-3f93ee19e346)

#### Writer benchmarks

We see a latency and memory usage improvement across all Writer configurations. As an example `ODataMessageWriter-Utf8JsonWriter-Async` has an ~8.5% improvement in latency and 24% drop in memory allocations.

```
BenchmarkDotNet=v0.13.1, OS=Windows 10.0.22621
Intel Xeon W-2123 CPU 3.60GHz, 1 CPU, 8 logical and 4 physical cores
.NET SDK=8.0.100
  [Host]     : .NET 6.0.25 (6.0.2523.51912), X64 RyuJIT
  DefaultJob : .NET 6.0.25 (6.0.2523.51912), X64 RyuJIT
```

**Before**
|           Method |                                           WriterName |       Mean |      Error |     StdDev |     Median |      Gen 0 |     Gen 1 |  Allocated |
|----------------- |----------------------------------------------------- |-----------:|-----------:|-----------:|-----------:|-----------:|----------:|-----------:|
| WriteToFileAsync |                                   ODataMessageWriter | 423.633 ms |  8.4293 ms | 16.2404 ms | 422.386 ms | 53000.0000 |         - | 224,018 KB |
| WriteToFileAsync |                             ODataMessageWriter-Async | 719.716 ms | 12.4249 ms | 11.6222 ms | 721.246 ms | 63000.0000 | 1000.0000 | 266,770 KB |
| WriteToFileAsync |                      ODataMessageWriter-NoValidation | 346.953 ms |  6.0102 ms |  5.3279 ms | 348.297 ms | 47000.0000 |         - | 200,579 KB |
| WriteToFileAsync |                ODataMessageWriter-NoValidation-Async | 657.755 ms | 12.9246 ms | 25.5119 ms | 646.122 ms | 57000.0000 | 1000.0000 | 243,329 KB |
| WriteToFileAsync |                    ODataMessageWriter-Utf8JsonWriter | 392.370 ms |  6.5777 ms |  8.7811 ms | 393.420 ms | 51000.0000 |         - | 217,340 KB |
| WriteToFileAsync |              ODataMessageWriter-Utf8JsonWriter-Async | 598.974 ms |  5.6653 ms |  4.4231 ms | 598.411 ms | 51000.0000 |         - | 218,139 KB |
| WriteToFileAsync |       ODataMessageWriter-Utf8JsonWriter-NoValidation | 331.546 ms |  6.6227 ms |  6.1949 ms | 329.425 ms | 46000.0000 |         - | 193,903 KB |
| WriteToFileAsync | ODataMessageWriter-Utf8JsonWriter-NoValidation-Async | 533.909 ms |  3.3615 ms |  3.1443 ms | 532.770 ms | 46000.0000 |         - | 194,698 KB |

**After**

|           Method |                                           WriterName |       Mean |      Error |     StdDev |     Median |      Gen 0 |     Gen 1 |  Allocated |
|----------------- |----------------------------------------------------- |-----------:|-----------:|-----------:|-----------:|-----------:|----------:|-----------:|
| WriteToFileAsync |                                   ODataMessageWriter | 330.500 ms |  3.0234 ms |  2.6801 ms | 330.384 ms | 40000.0000 |         - | 171,285 KB |
| WriteToFileAsync |                             ODataMessageWriter-Async | 623.021 ms | 11.8429 ms | 12.1618 ms | 619.559 ms | 50000.0000 | 1000.0000 | 214,036 KB |
| WriteToFileAsync |                      ODataMessageWriter-NoValidation | 284.599 ms |  5.1672 ms |  6.5349 ms | 282.055 ms | 40500.0000 |         - | 171,278 KB |
| WriteToFileAsync |                ODataMessageWriter-NoValidation-Async | 575.517 ms |  9.9837 ms | 16.4034 ms | 570.288 ms | 50000.0000 | 1000.0000 | 214,029 KB |
| WriteToFileAsync |                    ODataMessageWriter-Utf8JsonWriter | 314.829 ms |  2.1455 ms |  1.7916 ms | 314.157 ms | 39000.0000 |  500.0000 | 164,606 KB |
| WriteToFileAsync |              ODataMessageWriter-Utf8JsonWriter-Async | 547.834 ms |  4.6274 ms |  4.1021 ms | 548.000 ms | 39000.0000 |         - | 165,419 KB |
| WriteToFileAsync |       ODataMessageWriter-Utf8JsonWriter-NoValidation | 282.200 ms |  5.4841 ms |  6.0956 ms | 280.777 ms | 39000.0000 |         - | 164,609 KB |
| WriteToFileAsync | ODataMessageWriter-Utf8JsonWriter-NoValidation-Async | 500.182 ms |  9.9743 ms | 15.2319 ms | 495.599 ms | 39000.0000 | 1000.0000 | 165,398 KB |

#### Concurrent HTTP requests benchmarks

Using `bombardier` on the `TestServer`:

We see an ~8% latency improvement when using the async writer and about 11% when using the sync writer. The gap between sync and async is still large. There's still a lot of work to do :)

**Before**

(sync)

```
bombardier -l -c 50 -d 30s https://localhost:7120/customers/ODataMessageWriter-Utf8JsonWriter
```

```
Statistics        Avg      Stdev        Max
  Reqs/sec      1948.04    1871.73   24414.38
  Latency       27.83ms     6.96ms   243.49ms
  Latency Distribution
     50%    26.59ms
     75%    30.85ms
     90%    38.14ms
     95%    45.50ms
     99%    59.13ms
  HTTP codes:
    1xx - 0, 2xx - 53834, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    50.84MB/s
```

(async)

```
bombardier -l -c 50 -d 30s https://localhost:7120/customers/ODataMessageWriter-Utf8JsonWriter-Async
```
```
Statistics        Avg      Stdev        Max
  Reqs/sec      1467.15    1671.74   25448.70
  Latency       38.49ms     8.59ms   272.24ms
  Latency Distribution
     50%    36.15ms
     75%    46.81ms
     90%    54.53ms
     95%    63.47ms
     99%    80.15ms
  HTTP codes:
    1xx - 0, 2xx - 38960, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    36.76MB/s
```

**After**

(sync)
```
bombardier -l -c 50 -d 30s https://localhost:7120/customers/ODataMessageWriter-Utf8JsonWriter
```
```
bombardier -l -c 50 -d 30s https://localhost:7120/customers/ODataMessageWriter-Utf8JsonWriter
Bombarding https://localhost:7120/customers/ODataMessageWriter-Utf8JsonWriter for 30s using 50 connection(s)
[=================================================================================================================================================================================================================================================================================] 30s
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec      2044.23    1621.43   30253.21
  Latency       24.56ms     5.42ms   212.18ms
  Latency Distribution
     50%    23.70ms
     75%    31.45ms
     90%    35.39ms
     95%    40.41ms
     99%    52.06ms
  HTTP codes:
    1xx - 0, 2xx - 61100, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    57.64MB/s
```

(async)

```
bombardier -l -c 50 -d 30s https://localhost:7120/customers/ODataMessageWriter-Utf8JsonWriter-Async
```
```
Statistics        Avg      Stdev        Max
  Reqs/sec      1548.30    2831.05   89738.59
  Latency       35.64ms     6.90ms   217.80ms
  Latency Distribution
     50%    33.69ms
     75%    41.64ms
     90%    48.24ms
     95%    56.33ms
     99%    71.01ms
  HTTP codes:
    1xx - 0, 2xx - 42080, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    39.72MB/s
```



### Checklist (Uncheck if it is not completed)

- [x] *Test cases added*
- [x] *Build and test with one-click build and test script passed*